chore: pretty printer improvements (407/447 tests passing)
- Fix GRANT statement line breaking
- Fix CLUSTER stmt USING line breaking
- Fix REFRESH MATERIALIZED VIEW line breaking
- Fix ParamRef overflow (render as u32)
- Fix ResTarget AS line breaking
- Fix function call soft line breaks after/before parens
- Fix CREATE SEQUENCE options line breaking
- Fix ALTER TABLE SET EXPRESSION line breaking
- Fix CREATE FUNCTION name line breaking
- Add INCLUDE clause emission for UNIQUE, PRIMARY KEY, EXCLUDE constraints
- Fix CREATE TRANSFORM line breaking
- Fix DROP SUBSCRIPTION CASCADE/RESTRICT values
- Add NullTest and BoolExpr precedence info for parentheses
- Add comparison operator non-associativity handling

+++ crates/pgls_pretty_print/src/nodes/param_ref.rs
@@ -8,8 +8,10 @@ use crate::{
 pub(super) fn emit_param_ref(e: &mut EventEmitter, n: &ParamRef) {
     e.group_start(GroupKind::ParamRef);
 
-    // Emit $1, $2, etc.
-    e.token(TokenKind::IDENT(format!("${}", n.number)));
+    // Emit $1, $2, etc. - handle overflow as unsigned
+    // n.number is i32, but PostgreSQL param numbers are always positive
+    // When large numbers overflow, we need to interpret as unsigned
+    e.token(TokenKind::IDENT(format!("${}", n.number as u32)));
 
     e.group_end();
 }
